Rearrange fetchers lists
Repeated data requests are now very cheap thanks to caching. It makes
*Thumb fetchers useless (little work is still more expensive that no
work).

AmazonThumb is not affected by that, because there is no corresponding
fetcher for full book info.

This commit removes LivelibThumb because it's code has long been outdated.
FantlabThumb is still here, though not in use anymore - at least it works
and might be useful as documentation in future.